route entities -- move route related classes to package routes
