+++ domain/src/test/java/org/mobilitydata/gtfsvalidator/domain/entity/gtfs/RouteTypeTest.java
@@ -1,6 +1,7 @@
 package org.mobilitydata.gtfsvalidator.domain.entity.gtfs;
 
 import org.junit.jupiter.api.Test;
+import org.mobilitydata.gtfsvalidator.domain.entity.gtfs.routes.RouteType;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
